% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{acf.ctrm}
\alias{acf.ctrm}
\title{Autocorrelation function}
\usage{
\method{acf}{ctrm}(x, OCTRM = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{\link[CTRM]{ctrm}}}

\item{OCTRM}{If FALSE (default), each magnitude is matched with its preceding
interarrival time. If TRUE, each magnitude is matched with its
succeeding interarrival time.}

\item{...}{Additional arguments passed to \code{\link[stats]{acf}}}
}
\description{
Calculates and plots the autocorrelation function for the bivariate
time series of interarrival times and magnitudes.
}
\examples{
  library(magrittr)
  flares \%>\% ctrm() \%>\% thin(k=150) \%>\% acf()
}
