% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{acf}
\alias{acf}
\title{Autocorrelation function}
\usage{
acf(x, ...)
}
\arguments{
\item{x}{time series or ctrm object.}

\item{...}{Additional arguments passed to \code{\link[stats]{acf}} or
\code{\link[CTRM]{acf}}}
}
\description{
Autocorrelation function
}
