% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Bitcoin trading data}
\format{An object of class \code{zoo} of length 2043.}
\source{
\url{https://datamarket.com/data/set/4aq9/various-bitcoin-currency-statistics}
}
\usage{
bitcoin
}
\description{
A time series of trading volume of bitcoins,
measured in "Bitcoin days destroyed", from
2010-12-04 until 2016-07-07.
}
\keyword{datasets}
