% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coredata.ctrm}
\alias{coredata.ctrm}
\title{Extract event magnitudes}
\usage{
\method{coredata}{ctrm}(x, ...)
}
\arguments{
\item{x}{A \code{\link{ctrm}} object}

\item{...}{Additional arguments passed to future methods.}
}
\value{
A numeric vector of event magnitudes
}
\description{
Return the event magnitudes of a \code{ctrm} object.
}
