% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coredata}
\alias{coredata}
\alias{coredata.default}
\title{Extracting/Replacing the Core Data of Objects}
\usage{
coredata(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{further arguments passed to methods.}
}
\description{
Generic functions for extracting the core data contained in a (more complex) object and replacing it.
}
