% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTRM.R
\docType{package}
\name{CTRM-package}
\alias{CTRM}
\alias{CTRM-package}
\title{Continuous Time Random Maxima}
\description{
Fitting Continuous Time Random Maxima models to data.
Stability plots for the Mittag-Leffler and
Generalized Pareto distributions display the distribution parameters
and the quality of the fit.
}
\details{
More details here.
}
