% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ctrm.R
\name{ctrm}
\alias{ctrm}
\title{Create a CTRM object}
\usage{
ctrm(x, y = NULL)
}
\arguments{
\item{x}{Either
\itemize{
  \item a \code{data.frame} with two columns, or
  \item a vector; then y must be a vector of same length, or
  \item a \code{zoo} object.
}
The first component/column must contain the event times and be
of class "numeric", "Date" or "POSIXct";
the second component/column must contain the event magnitudes and
be of type "numeric".}
}
\value{
An object of class \code{'ctrm'} based on a time series of
        magnitudes.
}
\description{
The constructor for the class \code{'ctrm'} (Continuous Time Random
Maxima).
}
\examples{
times <- cumsum(MittagLeffleR::rml(n = 1000, tail = 0.8, scale = 5))
magnitudes <- rexp(n = 1000)
sim_ctrm <- ctrm(times, magnitudes)
}
\seealso{
\link{CTRM}
}
