% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostic Plots for Exceedance Times.}
\usage{
diagnostics(ctrm)
}
\arguments{
\item{ctrm}{An object of class \code{\link{ctrm}}}
}
\description{
Generate the following plots:
\enumerate{
  \item auto-correlation of waiting times
  \item auto-correlation of magnitudes
  \item cross-correlation of waiting times and magnitudes
  \item Mittag-Leffler QQ-plot for exceedance times
  \item Empirical compula (exceedances & exceedance times)
  \item Hill plot for exceedance times
}
}
