% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flares}
\alias{flares}
\title{Solar flare data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2488 rows and 2 columns.
}
\source{
The Hard X-Ray Burst Spectrometer (HXRBS) on the Solar Maximum
Mission, 14 February 1980 - 2 December 1989.
\url{https://umbra.nascom.nasa.gov/smm/hxrbs.html}
}
\usage{
flares
}
\description{
A time series of peak rates of solar eruptions.
}
\seealso{
Other bursty-data: 
\code{\link{bitcoin}},
\code{\link{seaquakes}}
}
\concept{bursty-data}
\keyword{datasets}
