% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{interarrival}
\alias{interarrival}
\title{Get inter-arrival times}
\usage{
interarrival(ctre)
}
\arguments{
\item{ctre}{The underlying ctre object}
}
\description{
Extract inter-arrival times of threshold crossings from a ctre object
}
\examples{
  library(magrittr)
  bitcoin \%>\% ctre() \%>\% thin(k=100) \%>\% interarrival
}
