% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{length.ctre}
\alias{length.ctre}
\title{Get length of underlying time series}
\usage{
\method{length}{ctre}(x)
}
\arguments{
\item{x}{The ctre object containing the time series}
}
\description{
Extract length of underlying time series.
}
\examples{
  library(magrittr)
  flares \%>\% ctre() \%>\% thin(k=200) \%>\% length()
}
