% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{length.ctrm}
\alias{length.ctrm}
\title{Get length of underlying time series}
\usage{
\method{length}{ctrm}(x)
}
\arguments{
\item{x}{The ctrm object containing the time series}
}
\description{
Extract length of underlying time series.
}
