% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{magnitudes}
\alias{magnitudes}
\title{Extract event magnitudes}
\usage{
magnitudes(ctre, ...)
}
\arguments{
\item{ctre}{A \code{\link{ctre}} object}

\item{...}{Additional arguments passed to future methods.}
}
\value{
A numeric vector of event magnitudes
}
\description{
Return the event magnitudes of a \code{ctre} object.
}
\examples{
  library(magrittr)
  bitcoin \%>\% ctre() \%>\% magnitudes()
}
