% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLestimates.R
\name{MLestimates}
\alias{MLestimates}
\title{Mittag-Leffler estimates for varying thresholds}
\usage{
MLestimates(ctrm, plot_me = TRUE, tail = NULL, scale = NULL)
}
\description{
For a range of thresholds, return the parameters of the Mittag-Leffler
distribution fitted to the threshold exceedance times.
}
\details{
If \code{plot_me = TRUE}, the estimates are returned invisibly.
}
