% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot.ctre}
\alias{plot.ctre}
\title{Plot a ctre object}
\usage{
\method{plot}{ctre}(x, p = 0.05, ...)
}
\arguments{
\item{x}{The ctre object whose time series should be plotted}

\item{p}{The fraction of magnitudes that exceed the threshold to be plotted}

\item{...}{Additional plotting parameters}
}
\description{
Plot a ctre object
}
\examples{
  library(magrittr)
  flares \%>\% ctre() \%>\% plot(p = 0.02, log = 'y')
}
