% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot.ctrm}
\alias{plot.ctrm}
\title{Plot a ctrm object}
\usage{
\method{plot}{ctrm}(x, p = 0.05, ...)
}
\arguments{
\item{x}{The ctrm object whose time series should be plotted}

\item{p}{The fraction of magnitudes that exceed the threshold to be plotted}

\item{...}{Additional plotting parameters}
}
\description{
Plot a ctrm object
}
