% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mrp}
\alias{plot.mrp}
\title{Stability plots for CTRMs.}
\usage{
\method{plot}{mrp}(mrp, what = c("data", "MLtail", "MLscale", "GPshape",
  "GPscale", "MLqq"), ...)
}
\arguments{
\item{mrp}{An object of class \code{mrp}, see \link{new_mrp}.}

\item{what}{What to plot.}

\item{...}{Additional parameters, see details.}
}
\description{
Generate plots of the data and of the parameter estimates of the
Mittag-Leffler and Generalized Pareto distributions vs the threshold
height.
}
\details{
More details here.
}
