% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{qqestplot_static}
\alias{qqestplot_static}
\title{Static QQ Plot estimator}
\usage{
qqestplot_static(data, top_k = NULL, plot_me = TRUE, ...)
}
\arguments{
\item{data}{A vector of data.}

\item{top_k}{Only use the top_k largest values (in the tail) for the plot}

\item{plot_me}{Should a plot be produced? If not, only the estimate is returned.}

\item{...}{Additional plotting arguments}
}
\value{
An estimate of the Pareto tail exponent (invisible if plotted).
}
\description{
Generates a static QQ plot for Pareto tail estimates
}
