% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quakes}
\alias{quakes}
\title{Coral Sea Earthquake Data}
\format{An object of class \code{data.frame} with 754 rows and 2 columns.}
\source{
\url{http://seismicportal.eu/}
}
\usage{
quakes
}
\description{
A time series of earthquake magnitudes measured within the bounding box
[166'E, 167'E] $\times$ [13.5'S, 12.4'S]
}
\keyword{datasets}
