% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/time.R
\name{time}
\alias{time}
\alias{time.ctrm}
\alias{time}
\alias{time.ctrm}
\title{Get event times}
\usage{
\method{time}{ctrm}(ctrm)

\method{time}{ctrm}(ctrm)
}
\arguments{
\item{ctrm}{A \code{\link{ctrm}} object.}

\item{ctrm}{A \code{\link{ctrm}} object.}
}
\value{
A numeric vector of event times, or a time series.

A numeric vector of event times, or a time series.
}
\description{
Extract event times from a \code{ctrm} object.

Extract event times from a \code{ctrm} object
}
