% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coredata.R
\name{coredata.ctrm}
\alias{coredata.ctrm}
\title{Extract event magnitudes}
\usage{
coredata.ctrm(ctrm)
}
\arguments{
\item{ctrm}{A \code{\link{ctrm}} object}
}
\value{
A numeric vector of event magnitudes
}
\description{
Return the event magnitudes of a \code{ctrm} object.
}
