% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{interarrival}
\alias{interarrival}
\title{Get inter-arrival times}
\usage{
interarrival(ctrm)
}
\arguments{
\item{ctrm}{The underlying ctrm object}
}
\description{
Extract inter-arrival times of threshold crossings from a ctrm object
}
\examples{
  library(magrittr)
  bitcoin \%>\% ctrm() \%>\% thin(k=100) \%>\% interarrival
}
