% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLestimates.R
\name{MLestimates}
\alias{MLestimates}
\title{Mittag-Leffler estimates for varying thresholds}
\usage{
MLestimates(ctrm, plot_me = TRUE, tail = NULL, scale = NULL)
}
\arguments{
\item{ctrm}{A \code{\link{ctrm}} object}
}
\value{
A \code{data.frame} of Mittag-Leffler parameter estimates,
        one row for each threshold, which is returned invisibly
        unless \code{plot_me = FALSE}.
}
\description{
For a range of thresholds, return the parameters of the Mittag-Leffler
distribution fitted to the threshold exceedance times.
}
\details{
If \code{plot_me = TRUE}, the estimates are returned invisibly.
}
