% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mlqqplot}
\alias{mlqqplot}
\title{Mittag-Leffler QQ Plot of threshold waiting times}
\usage{
mlqqplot(x, ...)
}
\arguments{
\item{x}{Object for which a QQ-Plot against the Mittag-Leffler
distribution is to be created.}

\item{...}{Additional plotting arguments}
}
\description{
Generates a QQ plot for assessing the fit of a Mittag-Leffler
distribution.
}
