% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlqqplot.R
\name{mlqqplot}
\alias{mlqqplot}
\title{Mittag-Leffler QQ Plot of threshold waiting times}
\usage{
mlqqplot(data, tail = 1, ...)
}
\description{
Generates a QQ plot for assessing the fit of a Mittag-Leffler
distribution.
}
