% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plot.ctrm}
\alias{plot.ctrm}
\alias{print.ctrm}
\title{Plot a CTRM}
\usage{
\method{print}{ctrm}(ctrm)
}
\arguments{
\item{ctrm}{An object of class \code{\link{ctrm}}.}

\item{p}{The fraction of largest magnitudes at which the threshold
is plotted.}

\item{...}{Additional parameters, see details.}
}
\description{
Generate a plot of the data highlighting a threshold, the threshold
exceedances and threshold exceedance times.
}
