% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqestplot.R
\name{qqestplot}
\alias{qqestplot}
\title{QQ Plot estimator of tail exponent}
\usage{
qqestplot(data, static = FALSE, conf.int = FALSE, k = NULL, ...)
}
\description{
Generates static and dynamic QQ plots for threshold crossing times.
}
