% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{thin}
\alias{thin}
\title{Apply a higher threshold to a CTRM process}
\usage{
thin(ctrm, k)
}
\arguments{
\item{ctrm}{A \code{\link{ctrm}} object.}

\item{k}{Discard all but the \code{k} largest magnitudes.}
}
\value{
A \code{\link{ctrm}} object, with fewer (k) observations.
}
\description{
Drop all but the \code{k} largest observations.
}
