% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{time}
\alias{time}
\alias{time}
\alias{time.ctre}
\title{Sampling Times of Time Series}
\usage{
time(x, ...)

\method{time}{ctre}(x, ...)
}
\arguments{
\item{x}{A time series or ctre object}

\item{...}{Extra arguments for future methods}
}
\value{
A numeric vector of event times, or a time series.
}
\description{
Sampling Times of Time Series

Extract event times from a \code{ctre} object.
}
\examples{
  library(magrittr)
  seaquakes \%>\% ctre() \%>\% thin(k=200) \%>\% time()
}
\seealso{
\code{\link{time.ctre}}
}
